duphold
this is an initial inclusion of duphold via a command-line flag.
it runs duphold serially on each sample. so, when run in cohort
mode with e.g. 20 samples and 20 threads, the duphold part will only
be able to use about 4 threads. For single (or < 4 samples) mode,
the duphold part will parallelize as expected up to ~4 threads.
this will improve in a future commit or version.